building interpreter with Ocaml
This project is to use Ocaml to build a simple simulator for general interpreter in IDE. Following the specific grammar rules of this project and input strings, the program can interpreter the information in string and compute the answer. This project will mainly use parsers to optimize the speed of computation(which is why this program has its own specific grammar). The program contains methods such as push, pop, log, swap, add, sub, mul, div, rem, neg, and, or, not, eq, less than, greater than, less or equal than, greater or equal than, let, ask, begin...end, if...else...end, call, throw, and lastly, try catch.